+++ src/state_manager/mod.rs
@@ -12,7 +12,7 @@ use anyhow::{bail, Context as _};
 use fil_actor_interface::init::{self, State};
 use rayon::prelude::ParallelBridge;
 pub use utils::is_valid_for_sending;
-mod vm_circ_supply;
+pub mod vm_circ_supply;
 pub use self::errors::*;
 use crate::beacon::BeaconSchedule;
 use crate::blocks::{Tipset, TipsetKeys};
@@ -21,7 +21,8 @@ use crate::chain::{
     ChainStore, HeadChange,
 };
 use crate::interpreter::{
-    resolve_to_key_addr, BlockMessages, CalledAt, ExecutionContext, IMPLICIT_MESSAGE_GAS_LIMIT, VM,
+    resolve_to_key_addr, ApplyResult, BlockMessages, CalledAt, ExecutionContext,
+    IMPLICIT_MESSAGE_GAS_LIMIT, VM,
 };
 use crate::message::{ChainMessage, Message as MessageTrait};
 use crate::networks::ChainConfig;
@@ -40,7 +41,7 @@ use chain_rand::ChainRand;
 use cid::Cid;
 
 use fil_actor_interface::miner::SectorOnChainInfo;
-use fil_actor_interface::miner::{MinerInfo, MinerPower};
+use fil_actor_interface::miner::{MinerInfo, MinerPower, Partition};
 use fil_actor_interface::*;
 use fil_actors_shared::fvm_ipld_amt::Amtv0 as Amt;
 use fil_actors_shared::fvm_ipld_bitfield::BitField;
@@ -59,7 +60,7 @@ use std::{num::NonZeroUsize, sync::Arc};
 use tokio::sync::{broadcast::error::RecvError, Mutex as TokioMutex, RwLock};
 use tracing::{debug, error, info, instrument, warn};
 use utils::structured;
-use vm_circ_supply::GenesisInfo;
+pub use vm_circ_supply::GenesisInfo;
 
 const DEFAULT_TIPSET_CACHE_SIZE: NonZeroUsize = nonzero!(1024usize);
 
@@ -422,7 +423,7 @@ where
                 epoch: height,
                 rand: Box::new(rand),
                 base_fee: tipset.blocks()[0].parent_base_fee().clone(),
-                circ_supply: genesis_info.get_circulating_supply(
+                circ_supply: genesis_info.get_vm_circulating_supply(
                     height,
                     &self.blockstore_owned(),
                     state_cid,
@@ -457,14 +458,14 @@ where
         let mut msg = msg.clone();
         msg.gas_limit = IMPLICIT_MESSAGE_GAS_LIMIT as u64;
 
-        let apply_ret = vm.apply_implicit_message(&msg)?;
+        let (apply_ret, duration) = vm.apply_implicit_message(&msg)?;
 
         Ok(ApiInvocResult {
             msg: msg.clone(),
             msg_rct: Some(apply_ret.msg_receipt()),
             msg_cid: msg.cid().map_err(|err| Error::Other(err.to_string()))?,
             error: apply_ret.failure_info().unwrap_or_default(),
-            duration: 0,
+            duration: duration.as_nanos().clamp(0, u64::MAX as u128) as u64,
             gas_cost: MessageGasCost::default(),
             execution_trace: structured::parse_events(apply_ret.exec_trace()).unwrap_or_default(),
         })
@@ -501,10 +502,9 @@ where
         // "next" tipset
         let epoch = ts.epoch() + 1;
         let genesis_info = GenesisInfo::from_chain_config(self.chain_config());
-
         // FVM requires a stack size of 64MiB. The alternative is to use `ThreadedExecutor` from
         // FVM, but that introduces some constraints, and possible deadlocks.
-        let ret = stacker::grow(64 << 20, || -> anyhow::Result<ApplyRet> {
+        let (ret, _) = stacker::grow(64 << 20, || -> ApplyResult {
             let mut vm = VM::new(
                 ExecutionContext {
                     heaviest_tipset: Arc::clone(&ts),
@@ -512,7 +512,7 @@ where
                     epoch,
                     rand: Box::new(chain_rand),
                     base_fee: ts.blocks()[0].parent_base_fee().clone(),
-                    circ_supply: genesis_info.get_circulating_supply(
+                    circ_supply: genesis_info.get_vm_circulating_supply(
                         epoch,
                         &self.blockstore_owned(),
                         &st,
@@ -1047,26 +1047,43 @@ where
         addr: &Address,
         ts: &Arc<Tipset>,
     ) -> Result<BitField, Error> {
+        self.all_partition_sectors(addr, ts, |partition| partition.faulty_sectors().clone())
+    }
+    /// Retrieves miner recoveries.
+    pub fn miner_recoveries(
+        self: &Arc<Self>,
+        addr: &Address,
+        ts: &Arc<Tipset>,
+    ) -> Result<BitField, Error> {
+        self.all_partition_sectors(addr, ts, |partition| partition.recovering_sectors().clone())
+    }
+
+    fn all_partition_sectors(
+        self: &Arc<Self>,
+        addr: &Address,
+        ts: &Arc<Tipset>,
+        get_sector: impl Fn(Partition<'_>) -> BitField,
+    ) -> Result<BitField, Error> {
         let actor = self
             .get_actor(addr, *ts.parent_state())?
             .ok_or_else(|| Error::State("Miner actor not found".to_string()))?;
 
         let state = miner::State::load(self.blockstore(), actor.code, actor.state)?;
 
-        let mut faults = Vec::new();
+        let mut partitions = Vec::new();
 
         state.for_each_deadline(
             &self.chain_config.policy,
             self.blockstore(),
             |_, deadline| {
                 deadline.for_each(self.blockstore(), |_, partition| {
-                    faults.push(partition.faulty_sectors().clone());
+                    partitions.push(get_sector(partition));
                     Ok(())
                 })
             },
         )?;
 
-        Ok(BitField::union(faults.iter()))
+        Ok(BitField::union(partitions.iter()))
     }
 
     /// Retrieves miner power.
@@ -1407,7 +1424,7 @@ where
     let genesis_info = GenesisInfo::from_chain_config(&chain_config);
     let create_vm = |state_root: Cid, epoch, timestamp| {
         let circulating_supply =
-            genesis_info.get_circulating_supply(epoch, &chain_index.db, &state_root)?;
+            genesis_info.get_vm_circulating_supply(epoch, &chain_index.db, &state_root)?;
         VM::new(
             ExecutionContext {
                 heaviest_tipset: Arc::clone(&tipset),