+++ CHANGELOG.md
@@ -29,6 +29,15 @@
 
 ### Added
 
+- [#3779](https://github.com/ChainSafe/forest/pull/3779) Implement the
+  `Filecoin.StateMinerRecoveries` lotus-compatible RPC API.
+- [#3745](https://github.com/ChainSafe/forest/pull/3745) Implement the
+  `Filecoin.StateCirculatingSupply` lotus-compatible RPC API.
+- [#3773](https://github.com/ChainSafe/forest/pull/3773) Implement the
+  `Filecoin.StateVMCirculatingSupplyInternal` lotus-compatible RPC API.
+- [#3748](https://github.com/ChainSafe/forest/pull/3748) Add timing for each
+  message and gas charge in the JSON output of
+  `forest-tool snapshot compute-state` and `Filecoin.StateCall` RPC API.
 - [#3720](https://github.com/ChainSafe/forest/pull/3750) Implement the
   `Filecoin.StateMinerInfo` lotus-compatible RPC API.